part-4/task-4.21: add: Blog List Expansions, step 9

+++ Part-4/blog/controllers/blogs.js
@@ -52,11 +52,22 @@ blogsRouter.put('/:id', async (request, response) => {
 });
 
 blogsRouter.delete('/:id', async (request, response) => {
-  const deletedBlog = await Blog.findByIdAndDelete(request.params.id);
+  const decodedToken = jwt.verify(request.token, process.env.SECRET);
+  if (!decodedToken.id) {
+    return response.status(401).json({ error: 'token invalid' });
+  }
 
-  if (!deletedBlog) {
+  const blog = await Blog.findById(request.params.id);
+
+  if (!blog) {
     return response.status(404).json({ error: 'Blog is not found' });
   }
+
+  if (blog.user.toString() !== decodedToken.id.toString()) {
+    return response.status(401).json({ error: 'blog can be deleted only by user who created it' });
+  }
+
+  await Blog.findByIdAndDelete(request.params.id);
 
   response.status(204).end();
 });
